Rename to make golint happy
don't use MixedCaps in package name; contenttypeMiddleware should be
contenttypemiddleware (golint)
